Change the checking for out of bounds coordinates
For a 5x5 table, the maximum value of the x and y coordinate should be 4
because we are using the Southwest most corner as the origin (0,0). We can't
put the robot on the edge of our table or else it will fall since the robot
has also length and width. Putting the robots center on the edge will tip
the robot off and fall from the table.